Adding additional tests
Adding tests where we need multiple expectations. Tried using
"withf" and closure but it did not work not sure why but suspect
that it had something todo with that it could not determine which
closure to use. Instead we switched to "with" predicate::eq. Using
the predicate::eq instead required that we use a struct wrapper
for the params that inherits the PartialEq trait.