Replace bakery profile with bakery.bashrc
Instead of using a bakery profile placed under /etc/profile.d
which will be called when calling a bash login we can use
the /etc/bash.bashrc which is called when calling a bash interactive.
We are also removing the ~/.bashrc volume to isolate the shell session
more. This might change in the future but for now let lock it down
a bit